added FLOAT definition for vectors and matrices

+++ src/common/face.h
@@ -12,9 +12,9 @@
 
 typedef struct {	/* a polygonal face */
 	FVECT  norm;		/* the plane's unit normal */
-	double  offset;		/* plane equation:  DOT(norm, v) == offset */
-	double  area;		/* area of face */
-	double  *va;		/* vertex array (o->oargs.farg) */
+	FLOAT  offset;		/* plane equation:  DOT(norm, v) == offset */
+	FLOAT  area;		/* area of face */
+	FLOAT  *va;		/* vertex array (o->oargs.farg) */
 	short  nv;		/* # of vertices */
 	short  ax;		/* axis closest to normal */
 } FACE;
